HW_6 Release

+++ Semester_1/Fq1jjeR/HW_6/3_Task.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    int N, M;
+    long long N, M;
 
     cout << "Введите количество N (строк): ";
     cin >> N;
